requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.3.2
cycler==0.12.1
fonttools==4.59.0
hydra-core==1.3.2
kiwisolver==1.4.8
laspy==2.6.1
laszip==0.2.3
matplotlib==3.10.3
numpy==2.3.2
omegaconf==2.3.0
packaging==25.0
pillow==11.3.0
progressbar2==4.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-utils==3.9.1
PyYAML==6.0.2
six==1.17.0
typing_extensions==4.14.1
